requirements.txt
guardrails-ai
python-dotenv
langchain
langchain-openai
llama-index
llama-index-llms-openai==0.2.16
llama-index-output-parsers-guardrails==0.2.0